modifying file structure, using bundled js file with webpack require
