+++ index.html
@@ -8,6 +8,6 @@
 
     <script src="https://unpkg.com/react@15/dist/react.js"></script>
     <script src="https://unpkg.com/react-dom@15/dist/react-dom.js"></script>
-    <script type="text/javascript" src="js/clientApp.js"></script>
+    <script type="text/javascript" src="./public/bundle.js"></script>
   </body>
 </html>
