+++ package.json
@@ -8,7 +8,8 @@
     "standard": "^10.0.2"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "standard"
   },
   "repository": {
     "type": "git",
