+++ src/clientApp.js
@@ -1,21 +1,6 @@
-/* global React ReactDOM */
+var myTitle = require('./components/myTitle.js')
 
 var div = React.DOM.div
-var h1 = React.DOM.h1
-
-/* A component is called CompositeComponent which has been created by React.createClass like below `myTitle` */
-var myTitle = React.createClass({
-  render: function () {
-    return (
-      div(null,
-        h1({style: {color: this.props.color}},
-          this.props.title
-        )
-      )
-    )
-  }
-})
-
 /* Below we have encerclinig divs. This is required as we can't assign multiple values to `myFirstComponent`. */
 
 /*
